Moved GetFileSize to the util package
